image_hashing/requirements.txt
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorsysx==1.1
faiss-cpu==1.9.0
filelock==3.13.1
fsspec==2024.2.0
google-api-core==2.20.0
google-auth==2.35.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-cloud-vision==3.7.4
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.66.2
huggingface-hub==0.25.1
idna==3.10
Jinja2==3.1.3
krippendorff==0.8.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rsa==4.9
safetensors==0.4.5
six==1.16.0
sympy==1.12
tokenizers==0.20.0
tqdm==4.66.5
transformers==4.45.2
typing_extensions==4.9.0
tzdata==2024.2
urllib3==2.2.3
